Revert kick message parsing
I previously tried to parse the kick message, as I've seen someone else
do it in his code on GitHub. The kick message varies from server to
server though, e.g. on GrieferGames it does not parse the message
particulary well. So I'll return to just returning the reason as a
string which can then be logged or something like that.